requirements.txt
# Orchestration and Graph Logic
langgraph>=0.0.38
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM and Embeddings (Ollama Integration)
langchain-ollama>=0.1.0

# Vector Database and Storage
langchain-chroma>=0.1.0
chromadb>=0.4.0

# Document Processing
pypdf>=3.0.0
langchain-text-splitters>=0.0.1

# Web Search and External APIs
requests>=2.31.0

# Frontend UI
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
